+++ packages/remirror__extension-codemirror6/src/index.ts
@@ -1,2 +1,2 @@
+export { CodeMirror6Extension } from './codemirror-extension';
 export type { CodeMirrorExtensionOptions } from './codemirror-types';
-export * from './codemirror6-extension';
